requirements.txt
openai
pillow
moviepy
python-dotenv
pydantic
tqdm